nlp/requirements.txt
fastapi==0.95.2
uvicorn==0.22.0
pydantic==1.10.8
spacy==3.5.3
numpy==1.24.3
python-dotenv==1.0.0
pytest==7.3.1
httpx==0.24.1
presidio-analyzer==2.2.32
presidio-anonymizer==2.2.32 